services/backend/requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
GitPython
python-dotenv
tiktoken
pytest
